+++ iop/usb/keyboard/src/ps2kbd.c
@@ -492,7 +492,7 @@ void ps2kbd_getkeys(u8 keyMods, u8 ledStatus, const u8 *keys, kbd_dev *dev)
       currChars[0] = 0;
       currChars[1] = 0;
 
-      if(lineEndP == (tempPos - 1))
+      if(lineEndP == (u32)(tempPos - 1))
 	{
 	  break;
 	}
@@ -570,7 +570,7 @@ void ps2kbd_getkeys(u8 keyMods, u8 ledStatus, const u8 *keys, kbd_dev *dev)
 
       if((currChars[0] == PS2KBD_ESCAPE_KEY) && (currChars[1] != 0))
 	{
-	  if(lineEndP != (tempPos - 2))
+	  if(lineEndP != (u32)(tempPos - 2))
 	    {
 	      lineBuffer[lineEndP++] = currChars[0];
 	      lineEndP %= lineSize;
@@ -635,7 +635,7 @@ void ps2kbd_getkeys_raw(u8 newKeyMods, u8 oldKeyMods, u8 *new, const u8 *old)
       int currMod = (1 << loopKey);
       if(keyMods & currMod)
 	{
-	  if(lineEndP == (tempPos - 2))
+	  if(lineEndP == (u32)(tempPos - 2))
 	    {
 	      return;
 	    }
@@ -663,7 +663,7 @@ void ps2kbd_getkeys_raw(u8 newKeyMods, u8 oldKeyMods, u8 *new, const u8 *old)
 
   for(loopKey = 0; loopKey < PS2KBD_MAXKEYS; loopKey++)
     {
-      if(lineEndP == (tempPos - 2))
+      if(lineEndP == (u32)(tempPos - 2))
 	{
 	  return;
 	}
@@ -682,7 +682,7 @@ void ps2kbd_getkeys_raw(u8 newKeyMods, u8 oldKeyMods, u8 *new, const u8 *old)
 
   for(loopKey = 0; loopKey < PS2KBD_MAXKEYS; loopKey++)
     {
-      if(lineEndP == (tempPos - 2))
+      if(lineEndP == (u32)(tempPos - 2))
 	{
 	  return;
 	}
@@ -852,7 +852,7 @@ void ps2kbd_ioctl_setreadmode(u32 readmode)
 {
   int devLoop;
 
-  if(readmode == kbd_readmode) return;
+  if((int)readmode == kbd_readmode) return;
 
   if((readmode == PS2KBD_READMODE_NORMAL) || (readmode == PS2KBD_READMODE_RAW))
     {
@@ -1172,7 +1172,7 @@ void repeat_thread(void *arg)
 
 	      if((devices[devLoop]->repeatkeys[0]) && (devices[devLoop]->repeatkeys[1]))
 		{
-		  if(lineEndP != (tempPos - 2))
+		  if(lineEndP != (u32)(tempPos - 2))
 		    {
 		      lineBuffer[lineEndP++] = devices[devLoop]->repeatkeys[0];
 		      lineEndP %= lineSize;
@@ -1184,7 +1184,7 @@ void repeat_thread(void *arg)
 		}
 	      else if(devices[devLoop]->repeatkeys[0])
 		{
-		  if(lineEndP != (tempPos - 1))
+		  if(lineEndP != (u32)(tempPos - 1))
 		    {
 		      lineBuffer[lineEndP++] = devices[devLoop]->repeatkeys[0];
 		      lineEndP %= lineSize;
